move ChatInfo to tailwind
