+++ chat-react2/src/components/Home/Main/Main.tsx
@@ -5,7 +5,7 @@ import { IoMdAttach } from "react-icons/io";
 import { IoIosArrowDown } from "react-icons/io";
 import Message from "./Message/Message";
 import ChatHeader from "./ChatHeader";
-import ChatInfo from "./ChatInfo/ChatInfo";
+import ChatInfo from "./ChatInfo";
 import "./Main.scss";
 
 const Main = () => {
